+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Reading a test notification message off the SQS queue no longer causes the logs to show an error #7
 * Fixed a bug where the scale down actions where not triggering correctly, causing replication agents to never be removed
 * Fixed NoSuchElementException error when scaling publish up bug #4
+* Fixed unable to delete flush and replication agents #11
 
 ### 0.9.1
 * Initial version